Fix adding transactions always resulting in an error being thrown
The Mongoose ODM seems to always call the error callback passed into the .save method, but passing in "null" as the error if there was not one. As a result, an error is always thrown as the
callback previously always threw an error when called. To fix this, the callback has been modified to only throw an error if the "err" argument is not null.