Replace App.jsx code with routing logic
The code in the App component has been moved to a new component called "Home", while the App component is now used for client-side routing. "Home" is a more appropriate name for the
function of the App component, and the logic for React Router should be at the top level.