Modify BudgetForm to use the user state variable passed down from App
The isLoggedIn state variable has been removed due to it being unnecessary. The user state variable is no longer set locally and is passed down from the App component. When the user logs in,
the BudgetForm component sets the value of `user` to the entered name, which causes App to redirect the client to the `/:user` endpoint, which results in the Home component being rendered.